refactor: move serializer CustomChoicesField to utils
