+++ seed/serializers/analyses.py
@@ -7,22 +7,7 @@
 from rest_framework import serializers
 
 from seed.models import Analysis
-
-
-class CustomChoicesField(serializers.ChoiceField):
-    def to_representation(self, obj):
-        if obj == '' and self.allow_blank:
-            return obj
-        return self._choices[obj]
-
-    def to_internal_value(self, data):
-        if data == '' and self.allow_blank:
-            return ''
-
-        for key, val in self._choices.items():
-            if val == data:
-                return key
-        self.fail('invalid_choice', input=data)
+from seed.serializers.utils import CustomChoicesField
 
 
 class AnalysisSerializer(serializers.ModelSerializer):
